Implement Renderer::LensFlare::RenderSlot

+++ src/Renderer/Sprite.h
@@ -12,7 +12,7 @@ namespace Renderer
 
 		int16_t DrawScaled(int16_t xPos,
 			int16_t yPos,
-			int32_t sheetIndex,
+			int16_t sheetIndex,
 			int32_t tileIndex,
 			uint32_t red,
 			uint32_t green,
